# avx2: Eliminate allocs and clean up