--- a/md5-util_amd64.go
+++ b/md5-util_amd64.go
@@ -6,10 +6,6 @@
 
 package md5simd
 
-import (
-	"sort"
-)
-
 // Helper struct for sorting blocks based on length
 type lane struct {
 	len uint
@@ -25,15 +21,22 @@ type maskRounds struct {
 func generateMaskAndRounds8(input [8][]byte, mr *[8]maskRounds) (rounds int) {
 	// Sort on blocks length small to large
 	var sorted [8]lane
-	for c, inpt := range input {
+	for c, inpt := range input[:] {
 		sorted[c] = lane{uint(len(inpt)), uint(c)}
+		for i := c - 1; i >= 0; i-- {
+			// swap so largest is at the end...
+			if sorted[i].len > sorted[i+1].len {
+				sorted[i], sorted[i+1] = sorted[i+1], sorted[i]
+				continue
+			}
+			break
+		}
 	}
-	sort.Slice(sorted[:], func(i, j int) bool { return sorted[i].len < sorted[j].len })
 
 	// Create mask array including 'rounds' (of processing blocks of 64 bytes) between masks
 	m, round := uint64(0xff), uint64(0)
 
-	for _, s := range sorted {
+	for _, s := range sorted[:] {
 		if s.len > 0 {
 			if uint64(s.len)>>6 > round {
 				mr[rounds] = maskRounds{m, (uint64(s.len) >> 6) - round}
@@ -47,18 +50,24 @@ func generateMaskAndRounds8(input [8][]byte, mr *[8]maskRounds) (rounds int) {
 }
 
 func generateMaskAndRounds16(input [16][]byte, mr *[16]maskRounds) (rounds int) {
-
 	// Sort on blocks length small to large
 	var sorted [16]lane
-	for c, inpt := range input {
+	for c, inpt := range input[:] {
 		sorted[c] = lane{uint(len(inpt)), uint(c)}
+		for i := c - 1; i >= 0; i-- {
+			// swap so largest is at the end...
+			if sorted[i].len > sorted[i+1].len {
+				sorted[i], sorted[i+1] = sorted[i+1], sorted[i]
+				continue
+			}
+			break
+		}
 	}
-	sort.Slice(sorted[:], func(i, j int) bool { return sorted[i].len < sorted[j].len })
 
 	// Create mask array including 'rounds' (of processing blocks of 64 bytes) between masks
 	m, round := uint64(0xffff), uint64(0)
 
-	for _, s := range sorted {
+	for _, s := range sorted[:] {
 		if s.len > 0 {
 			if uint64(s.len)>>6 > round {
 				mr[rounds] = maskRounds{m, (uint64(s.len) >> 6) - round}