Eliminate allocations from avx2 code and clean up.

Before:
```
BenchmarkAvx2
BenchmarkAvx2/32KB-32               4285      265577 ns/op  1974.15 MB/s      3804 B/op       35 allocs/op
BenchmarkAvx2/64KB-32               3076      406067 ns/op  2582.27 MB/s      4866 B/op       47 allocs/op
BenchmarkAvx2/128KB-32              1500      767535 ns/op  2732.32 MB/s      7645 B/op       76 allocs/op
BenchmarkAvx2/256KB-32               868     1452767 ns/op  2887.11 MB/s     12381 B/op      126 allocs/op
BenchmarkAvx2/512KB-32               420     2759526 ns/op  3039.87 MB/s     21584 B/op      225 allocs/op
BenchmarkAvx2/1MB-32                 219     5465753 ns/op  3069.52 MB/s     40439 B/op      427 allocs/op
BenchmarkAvx2/2MB-32                 100    10769977 ns/op  3115.55 MB/s     77765 B/op      828 allocs/op
BenchmarkAvx2/4MB-32                  56    21607114 ns/op  3105.87 MB/s    150965 B/op     1621 allocs/op
BenchmarkAvx2/8MB-32                  28    42627404 ns/op  3148.63 MB/s    296721 B/op     3202 allocs/op
PASS
```

After:
```
BenchmarkAvx2/32KB-32               5714      207211 ns/op  2530.22 MB/s        42 B/op        1 allocs/op
BenchmarkAvx2/64KB-32               3750      309867 ns/op  3383.96 MB/s        34 B/op        1 allocs/op
BenchmarkAvx2/128KB-32              1999      582792 ns/op  3598.46 MB/s        31 B/op        1 allocs/op
BenchmarkAvx2/256KB-32              1090     1093578 ns/op  3835.40 MB/s        44 B/op        1 allocs/op
BenchmarkAvx2/512KB-32               588     2069728 ns/op  4053.00 MB/s        50 B/op        1 allocs/op
BenchmarkAvx2/1MB-32                 291     4113401 ns/op  4078.67 MB/s        90 B/op        1 allocs/op
BenchmarkAvx2/2MB-32                 146     8212334 ns/op  4085.86 MB/s       119 B/op        1 allocs/op
BenchmarkAvx2/4MB-32                  74    16256754 ns/op  4128.06 MB/s       184 B/op        1 allocs/op
BenchmarkAvx2/8MB-32                  37    31891903 ns/op  4208.52 MB/s       281 B/op        2 allocs/op
```